Extends GCD and LCM to arbitrary number of parameters

+++ src/irvine/oeis/a124/A124316.java
@@ -22,6 +22,6 @@ public class A124316 extends AbstractSequence {
   @Override
   public Z next() {
     ++mN;
-    return Integers.SINGLETON.sumdiv(mN, d -> Functions.SIGMA1.z(Functions.GCD.l((long) d, mN / d)));
+    return Integers.SINGLETON.sumdiv(mN, d -> Functions.SIGMA1.z(Functions.GCD.l(d, mN / d)));
   }
 }
